requirements.txt
PyQt6>=6.5
pefile>=2023.2.7
joblib>=1.3
lightgbm>=4.0
yara-python>=4.3
